fix(phd-table): Fix logic to bubble event
Many times the clicked target is an inner element, and not the target
with delegate/trigger/capture on it. Traverse the dom until we find
the function to run, else cancel the event or else Aurelia will throw
an error. This method is still horrible and needs to be properly fixed